docker-kafka/authentication/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
gunicorn==20.1.0
django-extensions==3.1.1
confluent-kafka==1.7.0
